+++ playground/src/react-interactable/Animator.js
@@ -66,7 +66,6 @@ class PhysicsAnimator {
 
 		let dx = 0;
 		let {vx,vy} = physicsObject
-		// console.log( physicsObject )
 		
 		if ( Math.abs(vx) > ANIMATOR_PAUSE_ZERO_VELOCITY ) {
 			dx = deltaTime * vx;
@@ -79,7 +78,6 @@ class PhysicsAnimator {
 			hadMovement = true;
 		}
 		
-		// console.log( {dx, dy} )
 		View.animate( dx, dy )
 
 		let cfwnm = hadMovement ? 0 : this.consecutiveFramesWithNoMovement + 1