+++ changelog.md
@@ -1,3 +1,7 @@
+## v0.6.1
+* Updates the position and display of the animated component mimic the react-native behavior
+* Pass styles down to the animated component
+
 ## v0.6.0
 * ReactDom version working ok
 * Added dist files using bili